requirements.txt
fastapi
Pillow
numpy
opencv-python-headless
tensorflow
uvicorn
python-multipart
